services/api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.1"
description = "Reletino"
authors = [
    "Jonas Lindberg <jonas.lindberg@futi.no>",
    "Jorge Lewis <jorge.lewis@futi.no>",
]
readme = "README.md"
homepage = "https://releti.no"
repository = "https://github.com/startino/reletino"
documentation = "https://api.rest.futi.no/docs"
license = "BSL-1.1 (https://github.com/startino/reletino/blob/alpha/LICENSE)"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "*"
fastapi = "^0.110.0"
tenacity = "^8.2.3"
supabase = "^2.4.0"
aiohttp = "^3.9.4"
redis = "^5.0.2"
langchain = "*"
langchain-community = "*"
langchain-openai = "*"
praw = "^7.7.1"
markdown = "*"
pandas = "*"
python-dotenv = "^1.0.1"
diskcache = "^5.6.3"
pydantic = "*"
fastapi-cli = "*"
html2text = "^2024.2.26"
langgraph = "^0.2.43"
playwright = "^1.48.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
